[PureGo] Unify shared stream metadata and tighten transport docs
Move the table-name header and auth into a shared metadata.go with a streamContext helper both protocols call (Arrow uses the same table-name header, so it was never proto-specific). Wrap handshake errors once at the open boundary instead of at mixed layers. Correct rawStream/Stream docs: the handshake flow is shared, only record framing and the setup/readiness hooks differ. Rename auth.go to metadata.go.
